Chat With Postgres Database/requirements.txt
psycopg2-binary
tk
streamlit 
langchain
sqlalchemy 
langchain-groq 
python-dotenv
